Separate Copy Times Modal into its own component file
